use float instead of int for intermediate results

+++ src/main.cpp
@@ -47,18 +47,18 @@ uint8_t scaleSpeed(uint8_t sp, uint8_t dx, uint8_t dmax) {
     } else if (sp < 200) {
       f = 2;
     }
-    uint16_t ret = sp % 100;
-    ret = ret * f * dmax / dx;
+float ret = sp % 100;
+    ret = ret * f * ((float) dmax / dx);
     if (ret > 54) {
       ret = ret / 2;
       if (ret > 99) {
-        return min(uint16_t(ret / 4),uint16_t(99));
+        return min((ret / 4),99);
       } else {
         return 100 + ret;
       }
     } else {
       return 200 + ret;
-    }  
+    }
   }
 }
